+++ processors/baggage/baggagetrace/processor.go
@@ -30,6 +30,8 @@ var _ trace.SpanProcessor = (*SpanProcessor)(nil)
 // The Baggage span processor duplicates onto a span the attributes found
 // in Baggage in the parent context at the moment the span is started.
 // The passed filter determines which baggage members are added to the span.
+//
+// If filter is nil, all baggage members will be added.
 func New(filter Filter) *SpanProcessor {
 	return &SpanProcessor{
 		filter: filter,
@@ -38,8 +40,13 @@ func New(filter Filter) *SpanProcessor {
 
 // OnStart is called when a span is started and adds span attributes for baggage contents.
 func (processor SpanProcessor) OnStart(ctx context.Context, span trace.ReadWriteSpan) {
+	filter := processor.filter
+	if filter == nil {
+		filter = AllowAllMembers
+	}
+
 	for _, member := range otelbaggage.FromContext(ctx).Members() {
-		if processor.filter(member) {
+		if filter(member) {
 			span.SetAttributes(attribute.String(member.Key(), member.Value()))
 		}
 	}